pyproject.toml
[project]
name = "ai-employee"
version = "0.3.0"
description = "Personal AI Employee - Gold Tier (Hackathon 0)"
requires-python = ">=3.13"
dependencies = [
    "watchdog>=4.0.0",
    "playwright>=1.40.0",
    "mcp>=1.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]
